Recipe repo. Updated ingredient repo too to use integer

+++ MegansPantry/src/main/java/com/solread/meganspantry/repository/IngredientRepository.java
@@ -6,7 +6,7 @@ import com.solread.meganspantry.model.Ingredient;
 
 import java.util.List;
 
-public interface IngredientRepository extends CrudRepository<Ingredient, Long> {
+public interface IngredientRepository extends CrudRepository<Ingredient, Integer> {
 
     public List<Ingredient> findByVeganTrue();
     public List<Ingredient> findByVegetarianTrue();
